+++ tket/src/Predicates/include/Predicates/PassGenerators.hpp
@@ -19,6 +19,7 @@
 #include "Mapping/LexiRoute.hpp"
 #include "Mapping/RoutingMethod.hpp"
 #include "Transformations/ContextualReduction.hpp"
+#include "Transformations/Decomposition.hpp"
 #include "Transformations/PauliOptimisation.hpp"
 
 namespace tket {
@@ -118,6 +119,39 @@ PassPtr gen_user_defined_swap_decomp_pass(const Circuit& replacement_circ);
  * the input, but that maximises expected circuit fidelity
  */
 PassPtr KAKDecomposition(double cx_fidelity = 1.);
+
+/**
+ * @brief Decomposes each TK2 gate into two-qubit gates.
+ *
+ * We currently support CX, ZZMax and ZZPhase.
+ *
+ * If one or more gate fidelities are provided, the two-qubit gate
+ * type achieving the highest fidelity will be chosen for the
+ * decomposition, as measured using squared trace fidelity.
+ * If no fidelities are provided, the TK2 gates will be decomposed
+ * exactly using CX gates.
+ *
+ * All TK2(α, β, γ) gates must be normalised to the Weyl chamber, i.e.
+ * 0.5 ≥ 𝛼 ≥ 𝛽 ≥ |𝛾|.
+ *
+ * Gate fidelities are passed as keyword arguments to perform noise-aware
+ * decompositions. We currently support `CX_fidelity`, `ZZMax_fidelity` and
+ * `ZZPhase_fidelity`. If provided, the `CX` and `ZZMax` fidelities must be
+ * given by a single floating point fidelity. The `ZZPhase` fidelity is given as
+ * a lambda float -> float, mapping a ZZPhase angle parameter to its fidelity.
+ * These parameters will be used to return the optimal decomposition of each TK2
+ * gate, taking noise into consideration.
+ *
+ * If the TK2 angles are symbolic values, the decomposition will be exact
+ * (i.e. not noise-aware). It is not possible in general to obtain optimal
+ * decompositions for arbitrary symbolic parameters, so consider substituting
+ * for concrete values if possible.
+ *
+ * @param fid The two-qubit gate fidelities (optional).
+ * @return PassPtr
+ */
+PassPtr DecomposeTK2(const Transforms::TwoQbFidelities& fid);
+PassPtr DecomposeTK2();
 
 /**
  * Resynthesize and squash three-qubit interactions.